# wsfl-frontend/api-docs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "WSFL (Winter Series Fantasy League) API",
        "description": "API for managing fantasy running league competitions",
        "termsOfService": "https://www.example.com/terms",
        "contact": {
            "name": "API Support",
            "url": "https://www.example.com",
            "email": "your.email@example.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://choosealicense.com/licenses/mit/"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:9000",
            "description": "Development server"
        }
    ],
...
    "paths": {
...
        "/api/drafts": {
            "get": {
                "tags": [
                    "Draft"
                ],
                "summary": "Get all drafts",
                "description": "Retrieves a paginated list of all drafts in the system, with optional filtering by league and season",
                "operationId": "getAllDrafts",
                "parameters": [
                    {
                        "name": "leagueId",
                        "in": "query",
                        "description": "League ID to filter drafts",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "season",
                        "in": "query",
                        "description": "Season number to filter drafts",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Zero-based page index (0..N)",
                        "required": false,
                        "schema": {
                            "minimum": 0,
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "The size of the page to be returned",
                        "required": false,
                        "schema": {
                            "minimum": 1,
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved drafts",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageDraftSummaryDto"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "example": "Access denied"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Draft"
                ],
                "summary": "Create new draft",
                "description": "Creates a new draft for a league's season",
                "operationId": "createDraft",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DraftCreateDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Draft created successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/DraftSummaryDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "example": "Invalid input data"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not authorized to create drafts",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "example": "Access denied"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "League not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "example": "League not found with id: 1"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/drafts/{id}/start": {
            "post": {
                "tags": [
                    "Draft"
                ],
                "summary": "Start draft",
                "description": "Initiates the draft process for a specific draft",
                "operationId": "startDraft",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the draft to start",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Draft started successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/DraftSummaryDto"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not authorized to start drafts",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "example": "Access denied"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Draft not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "example": "Draft not found with id: 1"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/drafts/{id}/end": {
            "post": {
                "tags": [
                    "Draft"
                ],
                "summary": "End draft",
                "description": "Completes the draft process for a specific draft",
                "operationId": "endDraft",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the draft to end",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Not authorized to end drafts",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "example": "Access denied"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Draft ended successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/DraftSummaryDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Draft not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "example": "Draft not found with id: 1"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/drafts/{id}": {
            "get": {
                "tags": [
                    "Draft"
                ],
                "summary": "Get draft by ID",
                "description": "Retrieves detailed information about a specific draft",
                "operationId": "getDraft",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the draft to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Draft found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/DraftSummaryDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Draft not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "example": "Draft not found with id: 1"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "LeagueSummaryDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Unique identifier of the league",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string",
                        "description": "Name of the league"
                    },
                    "season": {
                        "type": "integer",
                        "description": "Current season number",
                        "format": "int32"
                    }
                },
                "description": "Summary information about a league"
            },
...
            "DraftCreateDto": {
                "required": [
                    "leagueId",
                    "numberOfRounds",
                    "season",
                    "snakeOrder",
                    "startTime"
                ],
                "type": "object",
                "properties": {
                    "leagueId": {
                        "type": "integer",
                        "description": "ID of the league this draft belongs to",
                        "format": "int64",
                        "example": 1
                    },
                    "season": {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Season number for this draft",
                        "format": "int32",
                        "example": 2023
                    },
                    "numberOfRounds": {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Number of rounds in the draft",
                        "format": "int32",
                        "example": 15
                    },
                    "snakeOrder": {
                        "type": "boolean",
                        "description": "Whether the draft order should snake (reverse) each round",
                        "example": true
                    },
                    "startTime": {
                        "type": "string",
                        "description": "Scheduled start time for the draft",
                        "format": "date-time"
                    }
                },
                "description": "Draft creation data"
            },
            "DraftSummaryDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "Unique identifier of the draft",
                        "format": "int64"
                    },
                    "league": {
                        "$ref": "#/components/schemas/LeagueSummaryDto"
                    },
                    "season": {
                        "type": "integer",
                        "description": "Season number",
                        "format": "int32"
                    },
                    "numberOfRounds": {
                        "type": "integer",
                        "description": "Number of rounds in the draft",
                        "format": "int32"
                    },
                    "snakeOrder": {
                        "type": "boolean",
                        "description": "Whether the draft order snakes (reverses) each round"
                    },
                    "startTime": {
                        "type": "string",
                        "description": "Start time of the draft",
                        "format": "date-time"
                    },
                    "isComplete": {
                        "type": "boolean",
                        "description": "Whether the draft is complete"
                    },
                    "currentRound": {
                        "type": "integer",
                        "description": "Current round number",
                        "format": "int32"
                    },
                    "currentPick": {
                        "type": "integer",
                        "description": "Current pick number within the round",
                        "format": "int32"
                    },
                    "draftOrder": {
                        "type": "array",
                        "description": "List of team IDs in draft order",
                        "items": {
                            "type": "integer",
                            "description": "List of team IDs in draft order",
                            "format": "int64"
                        }
                    }
                },
                "description": "Summary information about a draft"
            },
...
            "PageableObject": {
                "type": "object",
                "properties": {
                    "offset": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "sort": {
                        "$ref": "#/components/schemas/SortObject"
                    },
                    "pageNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "pageSize": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "paged": {
                        "type": "boolean"
                    },
                    "unpaged": {
                        "type": "boolean"
                    }
                }
            },
            "SortObject": {
                "type": "object",
                "properties": {
                    "empty": {
                        "type": "boolean"
                    },
                    "sorted": {
                        "type": "boolean"
                    },
                    "unsorted": {
                        "type": "boolean"
                    }
                }
            },
...
            "PageDraftSummaryDto": {
                "type": "object",
                "properties": {
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DraftSummaryDto"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "sort": {
                        "$ref": "#/components/schemas/SortObject"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "last": {
                        "type": "boolean"
                    },
                    "numberOfElements": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "pageable": {
                        "$ref": "#/components/schemas/PageableObject"
                    },
                    "empty": {
                        "type": "boolean"
                    }
                }
            }
        }
    }
}